+++ code/src/Core/Naming/ValidationErrorType.cs
