+++ code/src/Core/Naming/ValidationResult.cs
