+++ code/src/Core/PostActions/Catalog/NewItemGeneration/CopyFilesToProjectPostAction.cs
@@ -9,6 +9,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using Microsoft.Templates.Core.Gen;
+using Microsoft.Templates.Core.Helpers;
 
 namespace Microsoft.Templates.Core.PostActions.Catalog
 {
