+++ code/src/Core/Templates/MetadataInfo.cs
