+++ code/src/Core/Templates/MetadataLocalizedInfo.cs
