+++ code/src/Core/Templates/MetadataType.cs
