+++ code/test/Core.Test/PostActions/Catalog/GenerateTestCertificatePostActionTest.cs
@@ -8,6 +8,7 @@ using System.IO;
 using System.Threading.Tasks;
 using Microsoft.TemplateEngine.Abstractions;
 using Microsoft.Templates.Core.Gen;
+using Microsoft.Templates.Core.Helpers;
 using Microsoft.Templates.Core.Locations;
 using Microsoft.Templates.Core.PostActions.Catalog;
 using Microsoft.Templates.Fakes;
