+++ code/tools/WtsPackagingTool/RemoteSourceWorker.cs
@@ -10,6 +10,7 @@ using System.Text;
 using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 using Microsoft.Templates.Core;
+using Microsoft.Templates.Core.Helpers;
 using Microsoft.Templates.Core.Locations;
 using Microsoft.WindowsAzure.Storage.Blob;
 using Newtonsoft.Json;
